test(settings): four subjects, four files, and the fixtures in the kit (#4938)
The suite #1135 named — `settings.test.tsx` at 2057 lines — was split
along the three subjects that ticket lists, and it stands at 746 lines
today. The entry-visibility half landed at 1604, so the defect moved
rather than closed: `settings-nav.test.tsx` is the file over the ceiling
now, and its own doc comment names a subject that is only half of what
it holds.

Split along what each part actually claims:

- `settings-nav.test.tsx` — WHICH pages the level offers, as a whole
  list: the shape, the floor a grantless principal reaches, and the
  seeded roles.
- `settings-grants.test.tsx` — ONE page, ONE grant: each entry's own
  requirement, asserted with the grant and without it.
- `settings-chrome.test.tsx` — what the chrome draws around a page: the
  rail's shape, the scope a page publishes, the rail's search.
- `settings-reach.test.tsx` — addressing a page the rail did not offer:
  the boundary, an address naming no page, and a page that is the
  reader's to read while the rail leaves it out.

The fixtures three of those share — the nav backend, the catalog-derived
label and page-set helpers, and the two settle helpers — move to
`settings.testkit.tsx`, which already declares itself the home for what
every `settings*.test.tsx` needs. The ones only the whole-list suite uses
stay with it.

Three doc comments the earlier split stranded are back above the block
they describe: the access-boundary paragraph sat sixty lines from its
describe, and the search and scope paragraphs had merged into one.

No case changed and none was dropped: 57 before, 57 after, and one
mutation to `visibleSettingsPages` kills 41 of them across all four
files.

Closes #1135
